requirements.txt
requests
dnspython
googlesearch-python
pandas